refactor(tools): 将工具/MCP 设置从弹窗迁移为独立页面
