+++ src/renderer/pages/settings/tools/components/McpManagementSection.tsx
@@ -1,0 +1,231 @@
+import { Button, Dropdown, Menu, Modal } from '@arco-design/web-react';
+import { Plus } from '@icon-park/react';
+import React, { useCallback, useEffect, useState } from 'react';
+import { useTranslation } from 'react-i18next';
+import { acpConversation } from '@/common/ipcBridge';
+import type { IMcpServer } from '@/common/storage';
+import AionScrollArea from '@/renderer/components/base/AionScrollArea';
+import { useMcpServers, useMcpAgentStatus, useMcpOperations, useMcpConnection, useMcpModal, useMcpServerCRUD, useMcpOAuth } from '@/renderer/hooks/mcp';
+import AddMcpServerModal from '../../components/AddMcpServerModal';
+import McpServerItem from '../../McpManagement/McpServerItem';
+import type { McpImportMode, MessageInstance } from '../types';
+
+export default function McpManagementSection({ message }: IMcpManagementSectionProps) {
+  const { t } = useTranslation();
+  const { mcpServers, extensionMcpServers, saveMcpServers } = useMcpServers();
+  const { agentInstallStatus, setAgentInstallStatus, isServerLoading, checkSingleServerInstallStatus } = useMcpAgentStatus();
+  const { syncMcpToAgents, removeMcpFromAgents } = useMcpOperations(mcpServers, message);
+  const { oauthStatus, loggingIn, checkOAuthStatus, login } = useMcpOAuth();
+
+  const onAuthRequired = useCallback(
+    (server: IMcpServer) => {
+      void checkOAuthStatus(server);
+    },
+    [checkOAuthStatus]
+  );
+
+  const { testingServers, handleTestMcpConnection } = useMcpConnection(mcpServers, saveMcpServers, message, onAuthRequired);
+  const { showMcpModal, editingMcpServer, deleteConfirmVisible, serverToDelete, mcpCollapseKey, showAddMcpModal, showEditMcpModal, hideMcpModal, showDeleteConfirm, hideDeleteConfirm, toggleServerCollapse } = useMcpModal();
+  const { handleAddMcpServer, handleBatchImportMcpServers, handleEditMcpServer, handleDeleteMcpServer, handleToggleMcpServer } = useMcpServerCRUD(mcpServers, saveMcpServers, syncMcpToAgents, removeMcpFromAgents, checkSingleServerInstallStatus, setAgentInstallStatus, message);
+
+  const [detectedAgents, setDetectedAgents] = useState<Array<{ backend: string; name: string }>>([]);
+  const [importMode, setImportMode] = useState<McpImportMode>('json');
+
+  useEffect(() => {
+    const loadAgents = async () => {
+      try {
+        const response = await acpConversation.getAvailableAgents.invoke();
+        if (response.success && response.data) {
+          setDetectedAgents(response.data.map((agent) => ({ backend: agent.backend, name: agent.name })));
+        }
+      } catch (error) {
+        console.error('Failed to load agents:', error);
+      }
+    };
+    void loadAgents();
+  }, []);
+
+  useEffect(() => {
+    const httpServers = mcpServers.filter((server) => server.transport.type === 'http' || server.transport.type === 'sse');
+    if (httpServers.length > 0) {
+      httpServers.forEach((server) => {
+        void checkOAuthStatus(server);
+      });
+    }
+  }, [mcpServers, checkOAuthStatus]);
+
+  const onOAuthLogin = useCallback(
+    async (server: IMcpServer) => {
+      const result = await login(server);
+
+      if (result.success) {
+        message.success(`${server.name}: ${t('settings.mcpOAuthLoginSuccess') || 'Login successful'}`);
+        void handleTestMcpConnection(server);
+      } else {
+        message.error(`${server.name}: ${result.error || t('settings.mcpOAuthLoginFailed') || 'Login failed'}`);
+      }
+    },
+    [login, message, t, handleTestMcpConnection]
+  );
+
+  const onAddMcpServer = useCallback(
+    async (serverData: Omit<IMcpServer, 'id' | 'createdAt' | 'updatedAt'>) => {
+      const addedServer = await handleAddMcpServer(serverData);
+      if (addedServer) {
+        void handleTestMcpConnection(addedServer);
+        if (addedServer.transport.type === 'http' || addedServer.transport.type === 'sse') {
+          void checkOAuthStatus(addedServer);
+        }
+        if (serverData.enabled) {
+          void syncMcpToAgents(addedServer, true);
+        }
+      }
+    },
+    [handleAddMcpServer, handleTestMcpConnection, checkOAuthStatus, syncMcpToAgents]
+  );
+
+  const onEditMcpServer = useCallback(
+    async (editingMcpServer: IMcpServer | undefined, serverData: Omit<IMcpServer, 'id' | 'createdAt' | 'updatedAt'>) => {
+      const updatedServer = await handleEditMcpServer(editingMcpServer, serverData);
+      if (updatedServer) {
+        void handleTestMcpConnection(updatedServer);
+        if (updatedServer.transport.type === 'http' || updatedServer.transport.type === 'sse') {
+          void checkOAuthStatus(updatedServer);
+        }
+        if (serverData.enabled) {
+          void syncMcpToAgents(updatedServer, true);
+        }
+      }
+    },
+    [handleEditMcpServer, handleTestMcpConnection, checkOAuthStatus, syncMcpToAgents]
+  );
+
+  const onBatchImportMcpServers = useCallback(
+    async (serversData: Omit<IMcpServer, 'id' | 'createdAt' | 'updatedAt'>[]) => {
+      const addedServers = await handleBatchImportMcpServers(serversData);
+      if (addedServers && addedServers.length > 0) {
+        addedServers.forEach((server) => {
+          void handleTestMcpConnection(server);
+          if (server.transport.type === 'http' || server.transport.type === 'sse') {
+            void checkOAuthStatus(server);
+          }
+          if (server.enabled) {
+            void syncMcpToAgents(server, true);
+          }
+        });
+      }
+    },
+    [handleBatchImportMcpServers, handleTestMcpConnection, checkOAuthStatus, syncMcpToAgents]
+  );
+
+  const onConfirmDelete = useCallback(async () => {
+    if (!serverToDelete) return;
+    hideDeleteConfirm();
+    await handleDeleteMcpServer(serverToDelete);
+  }, [serverToDelete, hideDeleteConfirm, handleDeleteMcpServer]);
+
+  const onAddJsonServer = useCallback(() => {
+    setImportMode('json');
+    showAddMcpModal();
+  }, [setImportMode, showAddMcpModal]);
+
+  const onAddOneClickServer = useCallback(() => {
+    setImportMode('oneclick');
+    showAddMcpModal();
+  }, [setImportMode, showAddMcpModal]);
+
+  const renderAddButton = () => {
+    if (detectedAgents.length > 0) {
+      return (
+        <Dropdown.Button
+          trigger='click'
+          type='outline'
+          position='br'
+          droplist={
+            <Menu>
+              <Menu.Item key='json' onClick={onAddJsonServer}>
+                {t('settings.mcpImportFromJSON')}
+              </Menu.Item>
+              <Menu.Item key='oneclick' onClick={onAddOneClickServer}>
+                {t('settings.mcpOneKeyImport')}
+              </Menu.Item>
+            </Menu>
+          }
+        >
+          <Plus fill='var(--primary)' size={18} />
+          {t('settings.mcpAddServer')}
+        </Dropdown.Button>
+      );
+    }
+
+    return (
+      <Button type='outline' icon={<Plus size={18} fill={'var(--primary)'} />} shape='round' onClick={onAddJsonServer}>
+        {t('settings.mcpAddServer')}
+      </Button>
+    );
+  };
+
+  return (
+    <div className='flex flex-col gap-16px min-h-0'>
+      <div className='flex gap-8px items-center justify-between'>
+        <div className='text-14px text-foreground'>{t('settings.mcpSettings')}</div>
+        <div>{renderAddButton()}</div>
+      </div>
+
+      <div className='flex-1 min-h-0'>
+        {mcpServers.length === 0 && extensionMcpServers.length === 0 ? (
+          <div className='py-24px text-center text-secondary text-14px border border-dashed rd-12px'>{t('settings.mcpNoServersFound')}</div>
+        ) : (
+          <AionScrollArea className={'max-h-360px max-h-none'} disableOverflow>
+            <div className='space-y-12px'>
+              {mcpServers.map((server) => (
+                <McpServerItem
+                  key={server.id}
+                  server={server}
+                  isCollapsed={mcpCollapseKey[server.id] || false}
+                  agentInstallStatus={agentInstallStatus}
+                  isServerLoading={isServerLoading}
+                  isTestingConnection={testingServers[server.id] || false}
+                  oauthStatus={oauthStatus[server.id]}
+                  isLoggingIn={loggingIn[server.id]}
+                  onToggleCollapse={() => toggleServerCollapse(server.id)}
+                  onTestConnection={handleTestMcpConnection}
+                  onEditServer={showEditMcpModal}
+                  onDeleteServer={showDeleteConfirm}
+                  onToggleServer={handleToggleMcpServer}
+                  onOAuthLogin={onOAuthLogin}
+                />
+              ))}
+              {extensionMcpServers.map((server) => (
+                <McpServerItem
+                  key={server.id}
+                  server={server}
+                  isCollapsed={mcpCollapseKey[server.id] || false}
+                  agentInstallStatus={agentInstallStatus}
+                  isServerLoading={isServerLoading}
+                  isTestingConnection={false}
+                  onToggleCollapse={() => toggleServerCollapse(server.id)}
+                  onTestConnection={handleTestMcpConnection}
+                  onEditServer={() => {}}
+                  onDeleteServer={() => {}}
+                  onToggleServer={() => Promise.resolve()}
+                  isReadOnly
+                />
+              ))}
+            </div>
+          </AionScrollArea>
+        )}
+      </div>
+
+      <AddMcpServerModal visible={showMcpModal} server={editingMcpServer} onCancel={hideMcpModal} onSubmit={editingMcpServer ? (serverData) => onEditMcpServer(editingMcpServer, serverData) : onAddMcpServer} onBatchImport={onBatchImportMcpServers} importMode={importMode} />
+
+      <Modal title={t('settings.mcpDeleteServer')} visible={deleteConfirmVisible} onCancel={hideDeleteConfirm} onOk={onConfirmDelete} okButtonProps={{ status: 'danger' }} okText={t('common.confirm')} cancelText={t('common.cancel')}>
+        <p>{t('settings.mcpDeleteConfirm')}</p>
+      </Modal>
+    </div>
+  );
+}
+
+interface IMcpManagementSectionProps {
+  message: MessageInstance;
+}
