+++ src/renderer/pages/settings/tools/index.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import PageWrapper from '@renderer/components/base/PageWrapper';
+import ToolsContent from './components/ToolsContent';
+
+export default function ToolsSettings() {
+  return (
+    <PageWrapper contentClassName='max-w-300'>
+      <ToolsContent />
+    </PageWrapper>
+  );
+}
